\name{ssBLAST-package}
\alias{ssBLAST-package}
\alias{ssBLAST}
\docType{package}
\title{\packageTitle{ssBLAST}}
\description{\packageDescription{ssBLAST}}
\section{Package Content}{\packageIndices{ssBLAST}}
\author{\packageAuthor{ssBLAST}}
\section{Maintainer}{\packageMaintainer{ssBLAST}}
\keyword{package}
