% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derep_sequences.r
\name{dereplicate_sequences}
\alias{dereplicate_sequences}
\title{Dereplicate reads in FASTQ file.}
\usage{
dereplicate_sequences(fastq_file_path, output_path = NULL)
}
\arguments{
\item{fastq_file_path}{String that either represents a path to a single fastq file, or a directory containing multiple fastq files.}

\item{output_path}{String for path that new files should be created. Default uses the `fastq_file_path`.}
}
\value{
fastq
}
\description{
Dereplicates read from a FASTQ file. Keeps quality score according to the first encounter of the sequence, and 
sequence ID from the corresponding read that owns the quality score.
}
