% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unaligned_BLAST_reads.r
\name{unaligned_BLAST}
\alias{unaligned_BLAST}
\title{Extract reads not aligned in BLAST}
\usage{
unaligned_BLAST(fastq_file_path, BLAST_file_path, output_path = NULL)
}
\arguments{
\item{fastq_file_path}{String that either represents a path to a single fastq file, or a directory containing multiple fastq files.}

\item{BLAST_file_path}{String that either represents a path to a single blast file, or a directory containing multiple blast files that correspond to `the fastq_file_path`.}

\item{output_path}{String for path that new files should be created. Default uses the `fastq_file_path`.}
}
\value{
fastq
}
\description{
Compares fastq file with corresponding BLAST file to find reads that were not found in BLAST search and writes to new file.
}
